backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
motor>=3.6.0
pymongo>=4.8.0
python-dotenv>=1.0.1
pydantic>=2.9.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
aiohttp>=3.10.0
numpy>=1.26.0
tensorflow>=2.15.0
opencv-python>=4.10.0
pillow>=10.0.0
cloudinary>=1.43.0
pydantic[email]>=2.9.0
